@@ -2,7 +2,7 @@
   "name": "@firestitch/truncate",
   "title": "Truncate",
   "description": "@firestitch/truncate",
-  "version": "7.0.9",
+  "version": "9.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/Firestitch/ngx-truncate"
@@ -28,43 +28,45 @@
     "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular-devkit/build-ng-packagr": "^0.13.8",
-    "@angular/animations": "^7.2.13",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.3.8",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
+    "@angular-devkit/build-angular": "~0.901.1",
+    "@angular-devkit/build-ng-packagr": "~0.901.1",
+    "@angular/animations": "^9.1.2",
+    "@angular/cdk": "^9.2.1",
+    "@angular/cli": "^9.1.1",
+    "@angular/common": "^9.1.2",
+    "@angular/compiler": "^9.1.2",
+    "@angular/compiler-cli": "^9.1.2",
+    "@angular/core": "^9.1.2",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "^9.1.2",
+    "@angular/material": "^9.2.1",
+    "@angular/platform-browser": "^9.1.2",
+    "@angular/platform-browser-dynamic": "^9.1.2",
+    "@angular/router": "^9.1.2",
     "@compodoc/compodoc": "^1.1.9",
-    "@firestitch/drawer": "^7.0.1",
-    "@firestitch/example": "^7.0.7",
-    "@firestitch/menu": "^7.0.1",
-    "@firestitch/message": "^7.1.4",
-    "codelyzer": "^5.0.0",
+    "@firestitch/drawer": "^8.3.3",
+    "@firestitch/example": "^9.0.0",
+    "@firestitch/iframe": "^8.0.0",
+    "@firestitch/menu": "^9.1.1",
+    "@firestitch/message": "^8.1.0",
+    "codelyzer": "^5.1.2",
     "cpx": "^1.5.0",
     "hammerjs": "^2.0.8",
-    "ng-packagr": "^5.1.0",
-    "ngx-toastr": "^10.0.2",
+    "lodash-es": "^4.17.15",
+    "ng-packagr": "^9.0.0",
+    "ngx-highlightjs": "^4.1.0-beta",
+    "ngx-toastr": "^12.0.1",
     "npm-run-all": "^4.1.5",
-    "rxjs": "^6.5.1",
-    "tsickle": "^0.34.3",
+    "rxjs": "^6.5.5",
     "tslint": "^5.16.0",
-    "typescript": "^3.2.4",
-    "webpack": "^4.30.0",
-    "webpack-dev-server": "^3.3.1",
-    "zone.js": "^0.9.0"
+    "typescript": "^3.8.3",
+    "zone.js": "~0.10.2"
   },
   "engines": {
     "node": ">=6.0.0"
   },
-  "peerDependencies": {}
+  "peerDependencies": {},
+  "dependencies": {
+    "tslib": "^1.10.0"
+  }
 }
